One more refactor to move common code into a utils module (is that the term in Python?)
