+++ parametric/test_headers_datadog.py
@@ -1,8 +1,9 @@
 import pytest
 
-from parametric.protos.apm_test_client_pb2 import DistributedHTTPHeaders
 from parametric.spec.trace import SAMPLING_PRIORITY_KEY, ORIGIN
 from parametric.spec.trace import span_has_no_parent
+from parametric.utils.headers import make_single_request_and_get_headers
+from parametric.utils.test_agent import get_span
 
 
 @pytest.mark.skip_library("golang", "not implemented")
@@ -12,19 +13,15 @@ def test_headers_datadog_extract_valid(test_agent, test_library):
     and activated properly.
     """
     with test_library:
-        distributed_message = DistributedHTTPHeaders()
-        distributed_message.http_headers["x-datadog-trace-id"] = "123456789"
-        distributed_message.http_headers["x-datadog-parent-id"] = "987654321"
-        distributed_message.http_headers["x-datadog-sampling-priority"] = "2"
-        distributed_message.http_headers["x-datadog-origin"] = "synthetics,=web"
-        distributed_message.http_headers["x-datadog-tags"] = "_dd.p.dm=-4"
+        headers = make_single_request_and_get_headers(test_library, [
+            ['x-datadog-trace-id', '123456789'],
+            ['x-datadog-parent-id', '987654321'],
+            ['x-datadog-sampling-priority', '2'],
+            ['x-datadog-origin', 'synthetics,=web'],
+            ['x-datadog-tags', '_dd.p.dm=-4'],
+        ])
 
-        with test_library.start_span(
-            name="name", service="service", resource="resource", http_headers=distributed_message
-        ) as span:
-            span.set_meta(key="http.status_code", val="200")
-
-    span = test_agent.wait_for_num_traces(num=1)[0][0]
+    span = get_span(test_agent)
     assert span.get("trace_id") == 123456789
     assert span.get("parent_id") == 987654321
     assert span["meta"].get(ORIGIN) == "synthetics,=web"
@@ -38,19 +35,15 @@ def test_headers_datadog_extract_invalid(test_agent, test_library):
     """Ensure that invalid Datadog distributed tracing headers are not extracted.
     """
     with test_library:
-        distributed_message = DistributedHTTPHeaders()
-        distributed_message.http_headers["x-datadog-trace-id"] = "0"
-        distributed_message.http_headers["x-datadog-parent-id"] = "0"
-        distributed_message.http_headers["x-datadog-sampling-priority"] = "2"
-        distributed_message.http_headers["x-datadog-origin"] = "synthetics"
-        distributed_message.http_headers["x-datadog-tags"] = "_dd.p.dm=-4"
+        headers = make_single_request_and_get_headers(test_library, [
+            ['x-datadog-trace-id', '0'],
+            ['x-datadog-parent-id', '0'],
+            ['x-datadog-sampling-priority', '2'],
+            ['x-datadog-origin', 'synthetics'],
+            ['x-datadog-tags', '_dd.p.dm=-4'],
+        ])
 
-        with test_library.start_span(
-            name="name", service="service", resource="resource", http_headers=distributed_message
-        ) as span:
-            span.set_meta(key="http.status_code", val="200")
-
-    span = test_agent.wait_for_num_traces(num=1)[0][0]
+    span = get_span(test_agent)
     assert span.get("trace_id") != 0
     assert span_has_no_parent(span)
     # assert span["meta"].get(ORIGIN) is None # TODO: Determine if we keep x-datadog-origin for an invalid trace-id/parent-id
@@ -64,9 +57,10 @@ def test_headers_datadog_inject_valid(test_agent, test_library):
     """Ensure that Datadog distributed tracing headers are injected properly.
     """
     with test_library:
-        with test_library.start_span(name="name") as span:
-            headers = test_library.inject_headers(span.span_id).http_headers.http_headers
-    span = test_agent.wait_for_num_traces(num=1)[0][0]
+        headers = make_single_request_and_get_headers(test_library, [
+        ])
+
+    span = get_span(test_agent)
     assert int(headers["x-datadog-trace-id"]) == span.get("trace_id")
     assert int(headers["x-datadog-parent-id"]) == span.get("span_id")
     assert int(headers["x-datadog-sampling-priority"]) == span["metrics"].get(SAMPLING_PRIORITY_KEY)
@@ -79,19 +73,15 @@ def test_headers_datadog_propagate_valid(test_agent, test_library):
     and injected properly.
     """
     with test_library:
-        distributed_message = DistributedHTTPHeaders()
-        distributed_message.http_headers["x-datadog-trace-id"] = "123456789"
-        distributed_message.http_headers["x-datadog-parent-id"] = "987654321"
-        distributed_message.http_headers["x-datadog-sampling-priority"] = "2"
-        distributed_message.http_headers["x-datadog-origin"] = "synthetics"
-        distributed_message.http_headers["x-datadog-tags"] = "_dd.p.dm=-4"
+        headers = make_single_request_and_get_headers(test_library, [
+            ['x-datadog-trace-id', '123456789'],
+            ['x-datadog-parent-id', '987654321'],
+            ['x-datadog-sampling-priority', '2'],
+            ['x-datadog-origin', 'synthetics'],
+            ['x-datadog-tags', '_dd.p.dm=-4'],
+        ])
 
-        with test_library.start_span(
-            name="name", service="service", resource="resource", http_headers=distributed_message
-        ) as span:
-            headers = test_library.inject_headers(span.span_id).http_headers.http_headers
-
-    span = test_agent.wait_for_num_traces(num=1)[0][0]
+    span = get_span(test_agent)
     assert headers["x-datadog-trace-id"] == "123456789"
     assert headers["x-datadog-parent-id"] != "987654321"
     assert headers["x-datadog-sampling-priority"] == "2"
@@ -106,26 +96,16 @@ def test_headers_datadog_propagate_invalid(test_agent, test_library):
     and the new span context is injected properly.
     """
     with test_library:
-        distributed_message = DistributedHTTPHeaders()
-        distributed_message.http_headers["x-datadog-trace-id"] = "0"
-        distributed_message.http_headers["x-datadog-parent-id"] = "0"
-        distributed_message.http_headers["x-datadog-sampling-priority"] = "2"
-        distributed_message.http_headers["x-datadog-origin"] = "synthetics"
-        distributed_message.http_headers["x-datadog-tags"] = "_dd.p.dm=-4"
+        headers = make_single_request_and_get_headers(test_library, [
+            ['x-datadog-trace-id', '0'],
+            ['x-datadog-parent-id', '0'],
+            ['x-datadog-sampling-priority', '2'],
+            ['x-datadog-origin', 'synthetics'],
+            ['x-datadog-tags', '_dd.p.dm=-4'],
+        ])
 
-        with test_library.start_span(
-            name="name", service="service", resource="resource", http_headers=distributed_message
-        ) as span:
-            headers = test_library.inject_headers(span.span_id).http_headers.http_headers
-
-    span = test_agent.wait_for_num_traces(num=1)[0][0]
     assert headers["x-datadog-trace-id"] != "0"
     assert headers["x-datadog-parent-id"] != "0"
     assert headers["x-datadog-sampling-priority"] != "2"
     # assert headers["x-datadog-origin"] == '' # TODO: Determine if we keep x-datadog-origin for an invalid trace-id/parent-id
     assert "_dd.p.dm=-4" not in headers["x-datadog-tags"]
-
-def get_span(test_agent):
-    traces = test_agent.traces()
-    span = traces[0][0]
-    return span
