+++ parametric/test_headers_tracestate_dd.py
@@ -2,9 +2,8 @@ from typing import Any
 
 import pytest
 
-from parametric.protos.apm_test_client_pb2 import DistributedHTTPHeaders
-from parametric.spec.trace import SAMPLING_PRIORITY_KEY, ORIGIN
 from parametric.spec.tracecontext import get_tracecontext
+from parametric.utils.headers import make_single_request_and_get_headers
 
 parametrize = pytest.mark.parametrize
 
@@ -414,19 +413,3 @@ def test_headers_tracestate_dd_propagate_propagatedtags(test_agent, test_library
     assert 't.dm:-4' in dd_items6
     assert 't.usr.id:baz64::' in dd_items6
     assert 't.new_key:val' in dd_items6
-
-def get_span(test_agent):
-    traces = test_agent.traces()
-    span = traces[0][0]
-    return span
-
-def make_single_request_and_get_headers(test_library, headers_list):
-    distributed_message = DistributedHTTPHeaders()
-    for key, value in headers_list:
-        distributed_message.http_headers[key] = value
-
-    with test_library.start_span(
-        name="name", service="service", resource="resource", http_headers=distributed_message
-    ) as span:
-        headers = test_library.inject_headers(span.span_id).http_headers.http_headers
-        return headers
