+++ parametric/utils/test_agent.py
@@ -1,0 +1,4 @@
+def get_span(test_agent):
+    traces = test_agent.traces()
+    span = traces[0][0]
+    return span
